backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
google-generativeai
sentence-transformers
networkx
numpy
supabase
python-dotenv
pydantic
